refactor main panel, updae layout and styles
