+++ src/components/MainPanel/MainPanel.module.css
@@ -4,6 +4,7 @@
   display: flex;
   flex-direction: column;
   height: 100%;
+  width: 100%;
   gap: 8px;
   padding: 12px;
   font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
