+++ src/sidepanel/App.module.css
@@ -1,0 +1,6 @@
+.panelWrapper {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
